backend/requirements.txt
requests==2.28.1
pyyaml==6.0
pycryptodome==3.15.0
Flask==2.3.1
pandas==2.0.1
Flask-SQLAlchemy==3.0.5
Flask-APScheduler==1.13.1
pymysql==1.1.0
langchain==0.0.321
openpyxl==3.1.2
pypdf==3.17.0
openai==0.28.1
zhipuai==1.0.7
atlassian-python-api==3.41.11
cryptography==42.0.7
celery==5.4.0
redis==5.0.8
flask_migrate
chardet